bug fixes and optimizations
